requirements.txt
binance-futures-connector
numpy
pandas
pandas-ta
plotly
kaleido